refactor(): remove custom elements
Put browser render tests in *index.js* as we only have these tests.
