--- src/helpers/identity.js
@@ -1,5 +1,0 @@
-function identity(x) {
-  return x
-}
-
-module.exports = identity
