+++ src/render-html.js
@@ -1,43 +1,6 @@
 let Rx = require(`@reactivex/rxjs`)
 let toHTML = require(`vdom-to-html`)
-let {replaceCustomElementsWithSomething, makeCustomElementsRegistry} =
-  require(`./custom-elements`)
-let {makeCustomElementInput, ALL_PROPS} = require(`./custom-element-widget`)
 let {transposeVTree} = require(`./transposition`)
-
-function makePropertiesDriverFromVTree(vtree) {
-  return {
-    get: (propertyName) => {
-      if (propertyName === ALL_PROPS) {
-        return Rx.Observable.of(vtree.properties)
-      } else {
-        return Rx.Observable.of(vtree.properties[propertyName])
-      }
-    },
-  }
-}
-
-function makeReplaceCustomElementsWithVTree$(CERegistry, driverName) {
-  return function replaceCustomElementsWithVTree$(vtree) {
-    return replaceCustomElementsWithSomething(vtree, CERegistry,
-      function toVTree$(_vtree, WidgetClass) {
-        let interactions = {get: () => Rx.Observable.empty()}
-        let props = makePropertiesDriverFromVTree(_vtree)
-        let input = makeCustomElementInput(interactions, props)
-        let output = WidgetClass.definitionFn(input)
-        let vtree$ = output[driverName].last()
-        /*eslint-disable no-use-before-define */
-        return convertCustomElementsToVTree(vtree$, CERegistry, driverName)
-        /*eslint-enable no-use-before-define */
-      })
-  }
-}
-
-function convertCustomElementsToVTree(vtree$, CERegistry, driverName) {
-  return vtree$
-    .map(makeReplaceCustomElementsWithVTree$(CERegistry, driverName))
-    .flatMap(transposeVTree)
-}
 
 function makeResponseGetter() {
   return function get(selector) {
@@ -63,12 +26,9 @@ function makeBogusSelect() {
   }
 }
 
-function makeHTMLDriver(customElementDefinitions = {}) {
-  let registry = makeCustomElementsRegistry(customElementDefinitions)
-  return function htmlDriver(vtree$, driverName) {
-    let vtreeLast$ = vtree$.last()
-    let output$ = convertCustomElementsToVTree(vtreeLast$, registry, driverName)
-      .map(vtree => toHTML(vtree))
+function makeHTMLDriver() {
+  return function htmlDriver(vtree$) {
+    let output$ = vtree$.switchMap(transposeVTree).last().map(toHTML)
     output$.get = makeResponseGetter()
     output$.select = makeBogusSelect()
     return output$
@@ -76,9 +36,5 @@ function makeHTMLDriver(customElementDefinitions = {}) {
 }
 
 module.exports = {
-  makePropertiesDriverFromVTree,
-  makeReplaceCustomElementsWithVTree$,
-  convertCustomElementsToVTree,
-
   makeHTMLDriver,
 }
